+++ resolwe/flow/views/entity.py
@@ -38,7 +38,7 @@ class EntityViewSet(CollectionViewSet):
     def get_queryset(self):  # pylint: disable=method-hidden
         """Return queryset."""
         if self.request and self.request.query_params.get('hydrate_data', False):
-            return self.queryset.prefetch_related('data__entity_set')
+            return self.queryset.prefetch_related('data__entity_set', 'data__collection_set')
 
         return self.queryset